FEAT: add safe bootstrap/update script.
This script is in the same spirit as ez_setup.py for pip. Instead of
trying to install/upgrade enstaller from itself, we use the egg (fetched
from S3 by default) to bootstrap itself. As we the code being updated is
different from the updating code, this is much simpler and more robust.