ENH: use a specific bdist_enegg for enthought-specific egg.
Enstaller internal version "logic" does not work when not using
<version>-<build>, so we really need an egg with that format and not the
standard setuptools <version_<python-version>.egg.

To avoid bad surprises, we move the enthought-specific egg code in a new
bdist_enegg command, which produce egg with filenames actually handled
by enstaller. We also modify the bootstrap script to use this format and
not the standard one.