ENH: upgrade now uses an out-of-process method + bootstrap script.
It is more reliable as we don't need to make the upgraded enstaller work
under the old (already imported) code.

As the bootstrap script is now included in the egg, we can also safely
update the bootstrap code without needing to make the bootstrap code
work for any version of enstaller.