FEAT: add PEP386WorkaroundVersion class.
This class encompasses NormalizedVersion-based comparison and
"string-based" comparison as used in enstaller.utils.comparable_version,
so that we don't accidentally try (and fail) comparing different types
when one operand could not be converted to NormalizedVersion.

It also ensures we don't drop trailing zeros in versions to keep
roundtripping through str, while ignoring them in comparison.