FEAT: add private property to access egg name.
The egg name is the non version part of the egg filename, which of
course had to be different than the name in our current index format.

E.g. MKL-10.3-1.egg has 'mkl' as a name, but 'MKL' as _egg_name. We need
the _egg_name as this is used for the name part of dependencies.